test(yara): split bridge test + relocate bulk + cargo fmt
Splits tests/yara_guard.rs into two surfaces:

1. crates/nab-yara-engine/tests/yara_guard.rs — bulk test suite,
   uses ONLY nab_yara_engine APIs (lives in the crate it tests)
2. tests/nab_security_bridge.rs (NEW) — the single test that bridges
   nab::security into the YARA path. Uses nab::security::fetch_yara
   re-exports (root crate has its own FetchGuardConfig copy in
   src/security/yara_engine.rs).

Plus cargo fmt the now-corrected files (cookies.rs + yara_guard.rs)
which were carrying pre-fix-PR-103 whitespace.

Both clippy + fmt + cargo build --tests are clean locally.

Closes the test (yara-engine) regression flagged on PR #104 by Codex.
